LibWeb/HTML: Correctly parse milliseconds in time inputs
This commit brings the way we parse time inputs closer to the spec by
no longer ignoring milliseconds in time strings.

+++ Libraries/LibWeb/HTML/Dates.h
@@ -43,7 +43,7 @@ struct YearMonthDay {
 struct HourMinuteSecond {
     i32 hour;
     i32 minute;
-    i32 second;
+    f32 second;
 };
 
 struct DateAndTime {
